gyro/requirements.txt
asgiref
autopep8
Django
pycodestyle
pytz
sqlparse
Unipath
dj-database-url
python-decouple
gunicorn
whitenoise
django-heroku
djangorestframework

